Split up task board page into multiple components
