+++ Client/src/components/domain/TaskBoard/DropIndicator.tsx
@@ -1,0 +1,18 @@
+import { StatusEnum } from "../../../Models/TaskEnums";
+import './TaskBoardViewPage.css';
+type DropIndicatorProps = {
+  beforeId: string | null;
+  column: StatusEnum;
+};
+
+const DropIndicator = ({ beforeId, column }: DropIndicatorProps) => {
+  return (
+    <div
+      data-before={beforeId || "-1"}
+      data-column={column}
+      className="drop-indicator"
+    />
+  );
+};
+
+export default DropIndicator;
